# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'What to watch REST API'
  description: "Данный проект представляет собой REST API для управления базой данных фильмов,\n                        включая функции оставления комментариев, регистрации пользователей и управления\n                        списком избранных фильмов. Сервис интегрируется с КиноПоиском, что позволяет\n                        получать комментарии к фильмам и предоставляет пользователям возможность активно взаимодействовать с ними."
  version: 1.0.0
servers:
  -
    url: 'https://localhost'
paths:
...
  '/api/v1/comments/{id}':
    patch:
      summary: ''
      operationId: patchApiV1CommentsId
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 19
                    text: texttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttext
                    rating: 10
                    created_at: '2024-09-25 22:06:36'
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 19
                      text:
                        type: string
                        example: texttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttext
                      rating:
                        type: integer
                        example: 10
                      created_at:
                        type: string
                        example: '2024-09-25 22:06:36'
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Запрашиваемая страница не существует'
                properties:
                  error:
                    type: string
                    example: 'Запрашиваемая страница не существует'
      tags:
        - Comment
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                film_id:
                  type: string
                  description: ''
                  example: null
                rating:
                  type: integer
                  description: 'This field is required when <code>comment_id</code> is <code>null</code>. Минимальное значение у поля value 1 Максимальное значение у поля value 10.'
                  example: 8
                text:
                  type: string
                  description: 'Поле value должно быть длиной минимум 50 символов.'
                  example: ejuwdvamsymueyjrgiynzubamjtylzvzoujskncepyshzygqzuepyfrmijsvol
                parent_id:
                  type: string
                  description: ''
                  example: null
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the comment.'
        example: 19
        required: true
        schema:
          type: integer
      -
        in: path
        name: comment_id
        description: 'ID комментария'
        example: 1
        required: true
        schema:
          type: integer
  '/api/v1/comments/{comment_id}':
    delete:
      summary: 'DELETE api/v1/comments/{comment_id}'
      operationId: dELETEApiv1commentscommentId
      description: "Пользователь может удалить свой комментарий, на который нет ответов.\n         Администратор может удалить любой комментарий, у которого нет ответов"
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Запрашиваемая страница не существует'
                properties:
                  error:
                    type: string
                    example: 'Запрашиваемая страница не существует'
      tags:
        - Comment
    parameters:
      -
        in: path
        name: comment_id
        description: 'ID комментария'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: null
security:
  -
    default: []